backend/requirements/development.txt
Django>=3.2,<5.0
Pillow>=10.1.0,<10.2.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.1,<5.4.0
psycopg2-binary>=2.9.9,<2.10.0
dj-database-url>=2.1.0,<2.2.0
django-debug-toolbar>=4.2.0,<4.3.0
django-redis>=5.4.0,<5.5.0


#code_formatter
autopep8  # https://github.com/hhatto/autopep8
flake8==6.0.0  # https://github.com/PyCQA/flake8
isort==5.12.0  # https://github.com/PyCQA/isort
black==23.1.0  # https://github.com/psf/black
pre-commit==3.0.4  # https://github.com/pre-commit/pre-commit
